api/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# AI Agent dependencies
openai>=1.0.0
langchain
langchain-core
langchain-openai
langgraph
mcp

# Additional dependencies
requests
beautifulsoup4

# AWS Services (optional)
boto3==1.34.0
opensearch-py==2.4.0
requests-aws4auth==1.2.3
